+++ baseapp/streaming.go
@@ -9,6 +9,7 @@ import (
 	abci "github.com/cometbft/cometbft/api/cometbft/abci/v1"
 	"github.com/spf13/cast"
 
+	"cosmossdk.io/log"
 	"cosmossdk.io/schema"
 	"cosmossdk.io/schema/appdata"
 	"cosmossdk.io/schema/decoding"
@@ -36,7 +37,7 @@ func (app *BaseApp) EnableIndexer(indexerOpts interface{}, keys map[string]*stor
 		Config:     indexerOpts,
 		Resolver:   decoding.ModuleSetDecoderResolver(appModules),
 		SyncSource: nil,
-		Logger:     app.logger.With("module", "indexer"),
+		Logger:     app.logger.With(log.ModuleKey, "indexer"),
 	})
 	if err != nil {
 		return err
@@ -163,14 +164,16 @@ func (p listenerWrapper) ListenFinalizeBlock(_ context.Context, req abci.Finaliz
 
 func (p listenerWrapper) ListenCommit(ctx context.Context, res abci.CommitResponse, changeSet []*storetypes.StoreKVPair) error {
 	if cb := p.listener.OnKVPair; cb != nil {
-		updates := make([]appdata.ModuleKVPairUpdate, len(changeSet))
+		updates := make([]appdata.ActorKVPairUpdate, len(changeSet))
 		for i, pair := range changeSet {
-			updates[i] = appdata.ModuleKVPairUpdate{
-				ModuleName: pair.StoreKey,
-				Update: schema.KVPairUpdate{
-					Key:    pair.Key,
-					Value:  pair.Value,
-					Delete: pair.Delete,
+			updates[i] = appdata.ActorKVPairUpdate{
+				Actor: []byte(pair.StoreKey),
+				StateChanges: []schema.KVPairUpdate{
+					{
+						Key:    pair.Key,
+						Value:  pair.Value,
+						Remove: pair.Delete,
+					},
 				},
 			}
 		}
@@ -181,9 +184,15 @@ func (p listenerWrapper) ListenCommit(ctx context.Context, res abci.CommitRespon
 	}
 
 	if p.listener.Commit != nil {
-		err := p.listener.Commit(appdata.CommitData{})
+		commitCb, err := p.listener.Commit(appdata.CommitData{})
 		if err != nil {
 			return err
+		}
+		if commitCb != nil {
+			err := commitCb()
+			if err != nil {
+				return err
+			}
 		}
 	}