+++ simapp/sim_bench_test.go
@@ -6,17 +6,18 @@ import (
 	"os"
 	"testing"
 
-	"cosmossdk.io/log"
-	"github.com/cosmos/cosmos-sdk/testutils/sims"
-
+	dbm "github.com/cosmos/cosmos-db"
 	flag "github.com/spf13/pflag"
 	"github.com/spf13/viper"
 	"github.com/stretchr/testify/require"
+
+	"cosmossdk.io/log"
 
 	"github.com/cosmos/cosmos-sdk/baseapp"
 	"github.com/cosmos/cosmos-sdk/client/flags"
 	"github.com/cosmos/cosmos-sdk/server"
 	simtestutil "github.com/cosmos/cosmos-sdk/testutil/sims"
+	"github.com/cosmos/cosmos-sdk/testutils/sims"
 	simtypes "github.com/cosmos/cosmos-sdk/types/simulation"
 	"github.com/cosmos/cosmos-sdk/x/simulation"
 	simcli "github.com/cosmos/cosmos-sdk/x/simulation/client/cli"
@@ -57,6 +58,9 @@ func BenchmarkFullAppSimulation(b *testing.B) {
 
 	app := NewSimApp(logger, db, nil, true, appOptions, interBlockCacheOpt(), baseapp.SetChainID(sims.SimAppChainID))
 
+	blockedAddrs, err := BlockedAddresses(app.InterfaceRegistry().SigningContext().AddressCodec())
+	require.NoError(b, err)
+
 	// run randomized simulation
 	simParams, simErr := simulation.SimulateFromSeedX(
 		b,
@@ -66,7 +70,7 @@ func BenchmarkFullAppSimulation(b *testing.B) {
 		simtestutil.AppStateFn(app.AppCodec(), app.AuthKeeper.AddressCodec(), app.StakingKeeper.ValidatorAddressCodec(), app.SimulationManager(), app.DefaultGenesis()),
 		simtypes.RandomAccounts,
 		simtestutil.SimulationOperations(app, app.AppCodec(), config, app.txConfig),
-		BlockedAddresses(),
+		blockedAddrs,
 		config,
 		app.AppCodec(),
 		app.txConfig.SigningContext().AddressCodec(),
@@ -83,6 +87,9 @@ func BenchmarkFullAppSimulation(b *testing.B) {
 	}
 
 	if config.Commit {
-		simtestutil.PrintStats(db)
+		db, ok := db.(dbm.DB)
+		if ok {
+			simtestutil.PrintStats(db)
+		}
 	}
 }