+++ simapp/simd/cmd/testnet_test.go
@@ -11,7 +11,6 @@ import (
 	corectx "cosmossdk.io/core/context"
 	"cosmossdk.io/depinject"
 	"cosmossdk.io/log"
-	"cosmossdk.io/x/auth"
 	banktypes "cosmossdk.io/x/bank/types"
 	"cosmossdk.io/x/staking"
 
@@ -21,6 +20,7 @@ import (
 	"github.com/cosmos/cosmos-sdk/testutil/configurator"
 	"github.com/cosmos/cosmos-sdk/types/module"
 	moduletestutil "github.com/cosmos/cosmos-sdk/types/module/testutil"
+	"github.com/cosmos/cosmos-sdk/x/auth"
 	genutiltest "github.com/cosmos/cosmos-sdk/x/genutil/client/testutil"
 	genutiltypes "github.com/cosmos/cosmos-sdk/x/genutil/types"
 )
@@ -45,7 +45,7 @@ func Test_TestnetCmd(t *testing.T) {
 	)
 	require.NoError(t, err)
 	require.NotNil(t, moduleManager)
-	require.Len(t, moduleManager.Modules, 8)
+	require.Len(t, moduleManager.Modules, 9) // the registered above + runtime
 
 	home := t.TempDir()
 	cdcOpts := codectestutil.CodecOptions{}