+++ store/cachekv/store_test.go
@@ -7,6 +7,7 @@ import (
 	dbm "github.com/cosmos/cosmos-db"
 	"github.com/stretchr/testify/require"
 
+	corestore "cosmossdk.io/core/store"
 	"cosmossdk.io/math/unsafe"
 	"cosmossdk.io/store/cachekv"
 	"cosmossdk.io/store/dbadapter"
@@ -465,7 +466,7 @@ func randInt(n int) int {
 }
 
 // useful for replaying a error case if we find one
-func doOp(t *testing.T, st types.CacheKVStore, truth dbm.DB, op int, args ...int) {
+func doOp(t *testing.T, st types.CacheKVStore, truth corestore.KVStoreWithBatch, op int, args ...int) {
 	t.Helper()
 	switch op {
 	case opSet:
@@ -491,7 +492,7 @@ func doOp(t *testing.T, st types.CacheKVStore, truth dbm.DB, op int, args ...int
 	}
 }
 
-func doRandomOp(t *testing.T, st types.CacheKVStore, truth dbm.DB, maxKey int) {
+func doRandomOp(t *testing.T, st types.CacheKVStore, truth corestore.KVStoreWithBatch, maxKey int) {
 	t.Helper()
 	r := randInt(totalOps)
 	switch r {
@@ -535,7 +536,7 @@ func assertIterateDomain(t *testing.T, st types.KVStore, expectedN int) {
 	require.NoError(t, itr.Close())
 }
 
-func assertIterateDomainCheck(t *testing.T, st types.KVStore, mem dbm.DB, r []keyRange) {
+func assertIterateDomainCheck(t *testing.T, st types.KVStore, mem corestore.KVStoreWithBatch, r []keyRange) {
 	t.Helper()
 	// iterate over each and check they match the other
 	itr := st.Iterator(nil, nil)
@@ -569,7 +570,7 @@ func assertIterateDomainCheck(t *testing.T, st types.KVStore, mem dbm.DB, r []ke
 	require.NoError(t, itr2.Close())
 }
 
-func assertIterateDomainCompare(t *testing.T, st types.KVStore, mem dbm.DB) {
+func assertIterateDomainCompare(t *testing.T, st types.KVStore, mem corestore.KVStoreWithBatch) {
 	t.Helper()
 	// iterate over each and check they match the other
 	itr := st.Iterator(nil, nil)
@@ -597,7 +598,7 @@ func checkIterators(t *testing.T, itr, itr2 types.Iterator) {
 
 //--------------------------------------------------------
 
-func setRange(t *testing.T, st types.KVStore, mem dbm.DB, start, end int) {
+func setRange(t *testing.T, st types.KVStore, mem corestore.KVStoreWithBatch, start, end int) {
 	t.Helper()
 	for i := start; i < end; i++ {
 		st.Set(keyFmt(i), valFmt(i))
@@ -606,7 +607,7 @@ func setRange(t *testing.T, st types.KVStore, mem dbm.DB, start, end int) {
 	}
 }
 
-func deleteRange(t *testing.T, st types.KVStore, mem dbm.DB, start, end int) {
+func deleteRange(t *testing.T, st types.KVStore, mem corestore.KVStoreWithBatch, start, end int) {
 	t.Helper()
 	for i := start; i < end; i++ {
 		st.Delete(keyFmt(i))