+++ testutil/x/counter/testutil/expected_keepers_mocks.go
@@ -7,7 +7,7 @@ package testutil
 import (
 	reflect "reflect"
 
-	types0 "cosmossdk.io/x/auth/types"
+	types0 "github.com/cosmos/cosmos-sdk/x/auth/types"
 
 	types "github.com/cosmos/cosmos-sdk/types"
 	gomock "github.com/golang/mock/gomock"